Ajuste dos relacionamentos entre as entidades

+++ aplicacao/gpa-extensao/src/main/java/ufc/quixada/npi/gpa/repository/AcaoExtensaoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import ufc.quixada.npi.gpa.model.AcaoExtensao;
 
 @Repository
-public interface AcaoExtensaoRepository extends CrudRepository<AcaoExtensao, Long> {
+public interface AcaoExtensaoRepository extends CrudRepository<AcaoExtensao, Integer> {
 
 }
