+++ aplicacao/gpa-extensao/src/main/java/ufc/quixada/npi/gpa/repository/DocumentoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import ufc.quixada.npi.gpa.model.Documento;
 
 @Repository
-public interface DocumentoRepository extends CrudRepository<Documento, Long> {
+public interface DocumentoRepository extends CrudRepository<Documento, Integer> {
 
 }
