+++ aplicacao/gpa-extensao/src/main/java/ufc/quixada/npi/gpa/repository/ParceriaExternaRepository.java
@@ -4,7 +4,8 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import ufc.quixada.npi.gpa.model.ParceriaExterna;
+
 @Repository
-public interface ParceriaExternaRepository extends CrudRepository<ParceriaExterna, Long>{
+public interface ParceriaExternaRepository extends CrudRepository<ParceriaExterna, Integer>{
 
 }
