+++ aplicacao/gpa-extensao/src/main/java/ufc/quixada/npi/gpa/repository/ParticipacaoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import ufc.quixada.npi.gpa.model.Participacao;
 
 @Repository
-public interface ParticipacaoRepository extends CrudRepository<Participacao, Long> {
+public interface ParticipacaoRepository extends CrudRepository<Participacao, Integer> {
 
 }
